# PERF: MultiIndex._engine use smaller dtypes