Use smaller dtypes in `MultiIndex._engine` if possible.
Reduce both loading time and memory (peak and used).
The improvement should be more relevant when working with big indices.
Below there are 2 examples.

- [ ] closes #xxxx (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

### Improvement when `engine.values` uses `uint32` instead of `uint64`:

- `idx._engine`:  from 8.53 s to 2.45 s
- `_engine.values.nbytes`: from 800000000 to 400000000
- `_engine`:
  - from: peak memory: 5805.98 MiB, increment: 5218.30 MiB
  - to: peak memory: 2794.02 MiB, increment: 2193.25 MiB


#### Time, Pandas 2.2.2:
```python
In [4]: %time idx = pd.MultiIndex.from_product([np.arange(1000), np.arange(1000), np.arange(100)], names=["x0", "x1", "x2"])
   ...: 
CPU times: user 250 ms, sys: 191 ms, total: 442 ms
Wall time: 457 ms

In [5]: %time idx._engine
   ...: 
CPU times: user 2.83 s, sys: 4.01 s, total: 6.84 s
Wall time: 8.53 s
Out[5]: <pandas.core.indexes.multi.MultiIndexUIntEngine at 0x117363d50>

In [6]: idx._engine.values.dtype
Out[6]: dtype('uint64')

In [7]: idx._engine.values.nbytes
Out[7]: 800000000
```

#### Time, PR branch:
```python
In [3]: %time idx = pd.MultiIndex.from_product([np.arange(1000), np.arange(1000), np.arange(100)], names=["x0", "x1", "x2"])
CPU times: user 282 ms, sys: 192 ms, total: 474 ms
Wall time: 565 ms

In [4]: %time idx._engine
CPU times: user 963 ms, sys: 838 ms, total: 1.8 s
Wall time: 2.45 s
Out[4]: <pandas.core.indexes.multi.MultiIndexUInt32Engine at 0x10ab54f90>

In [5]: idx._engine.values.dtype
Out[5]: dtype('uint32')

In [6]: idx._engine.values.nbytes
Out[6]: 400000000

```

#### Memory, Pandas 2.2.2:
```python
In [4]: %memit idx = pd.MultiIndex.from_product([np.arange(1000), np.arange(1000), np.arange(100)], names=["x0", "x1", "x2"])
peak memory: 587.48 MiB, increment: 477.49 MiB

In [5]: %memit idx._engine
peak memory: 5805.98 MiB, increment: 5218.30 MiB

```

#### Memory, PR branch:
```python
In [4]: %memit idx = pd.MultiIndex.from_product([np.arange(1000), np.arange(1000), np.arange(100)], names=["x0", "x1", "x2"])
peak memory: 600.49 MiB, increment: 477.43 MiB

In [5]: %memit idx._engine
peak memory: 2794.02 MiB, increment: 2193.25 MiB

```

### Improvement when `engine.values` still needs to use `uint64`:

- `idx._engine`:  from 17.6 s to 4.56 s
- `_engine.values.nbytes`: from 390625000 to 390625000 (unchanged since the dtype is the same)
- `_engine`:
  - from: peak memory: 9106.29 MiB, increment: 8472.42 MiB
  - to: peak memory: 5325.37 MiB, increment: 4674.05 MiB


#### Time, Pandas 2.2.2:
```python
In [3]: %time idx = pd.MultiIndex.from_product([np.arange(5)] * 11)
CPU times: user 265 ms, sys: 104 ms, total: 369 ms
Wall time: 371 ms

In [4]: %time idx._engine
CPU times: user 5.28 s, sys: 9.4 s, total: 14.7 s
Wall time: 17.6 s
Out[4]: <pandas.core.indexes.multi.MultiIndexUIntEngine at 0x11bde3cd0>

In [5]: idx._engine.values.dtype
Out[5]: dtype('uint64')

In [6]: idx._engine.values.nbytes
Out[6]: 390625000
```

#### Time, PR branch:
```python
In [3]: %time idx = pd.MultiIndex.from_product([np.arange(5)] * 11)
CPU times: user 268 ms, sys: 112 ms, total: 380 ms
Wall time: 381 ms

In [4]: %time idx._engine
CPU times: user 2.24 s, sys: 1.75 s, total: 3.99 s
Wall time: 4.56 s
Out[4]: <pandas.core.indexes.multi.MultiIndexUInt64Engine at 0x1209d0e40>

In [6]: idx._engine.values.dtype
Out[6]: dtype('uint64')

In [7]: idx._engine.values.nbytes
Out[7]: 390625000

```

#### Memory, Pandas 2.2.2:
```python
In [4]: %memit idx = pd.MultiIndex.from_product([np.arange(5)] * 11)
peak memory: 633.76 MiB, increment: 523.95 MiB

In [5]: %memit idx._engine
peak memory: 9106.29 MiB, increment: 8472.42 MiB

In [6]: idx._engine
Out[6]: <pandas.core.indexes.multi.MultiIndexUIntEngine at 0x117505750>

```

#### Memory, PR branch:
```python
In [4]: %memit idx = pd.MultiIndex.from_product([np.arange(5)] * 11)
peak memory: 651.32 MiB, increment: 523.31 MiB

In [5]: %memit idx._engine
peak memory: 5325.37 MiB, increment: 4674.05 MiB

In [6]: idx._engine
Out[6]: <pandas.core.indexes.multi.MultiIndexUInt64Engine at 0x117323290>

```